+++ test/src/tests-beaconUpload.ts
@@ -1,8 +1,8 @@
 import sinon from 'sinon';
 import fetchMock from 'fetch-mock/esm/client';
 import { expect } from 'chai';
-import { urls } from './config';
-import { apiKey, MPConfig, testMPID } from './config';
+import { urls } from './config/constants';
+import { apiKey, MPConfig, testMPID } from './config/constants';
 import { MParticleWebSDK } from '../../src/sdkRuntimeModels';
 import { event0 } from '../fixtures/events';
 import { batch1, batch2, batch3 } from '../fixtures/batches';
